+++ python-sdk/mapmodex/scripts/__init__.py
@@ -1,4 +1,4 @@
-from .custom_av2_map_converter_copy import create_av2_infos_mp
-from .custom_nusc_map_converter_copy import create_nuscenes_infos
+from .av2_reading import create_av2_infos_mp
+from .nuscenes_reading import create_nuscenes_infos
 from .peturbation import PerturbParameters
 from .data_reading import GetMapLayerGeom
