+++ python-sdk/mapmodex/scripts/av2_reading.py
