+++ python-sdk/mapmodex/scripts/modexmap_reading.py
