+++ python-sdk/mapmodex/scripts/nuscenes_reading.py
